all migrations updated

+++ database/migrations/2025_11_04_160854_create_availabilities_table.php
@@ -11,6 +11,10 @@ return new class extends Migration
      */
     public function up(): void
     {
+        if (Schema::hasTable('availabilities')) {
+            return;
+        }
+
         Schema::create('availabilities', function (Blueprint $table) {
             $table->id();
             $table->foreignId('user_id')->constrained()->onDelete('cascade');
